+++ SakuraTranslate/Models/PromptMessage.cs
@@ -1,0 +1,8 @@
+﻿namespace SakuraTranslate.Models
+{
+    public class PromptMessage
+    {
+        public string Role { get; set; }
+        public string Content { get; set; }
+    }
+}
