+++ SakuraTranslate/Models/TranslationModel.cs
@@ -1,0 +1,12 @@
+﻿namespace SakuraTranslate.Models
+{
+    public enum TranslationModel
+    {
+        SakuraV0_8,
+        SakuraV0_9,
+        SakuraV0_10,
+        SakuraV1_0,
+        Sakura32bV0_10,
+        GalTranslV2_6
+    }
+}
