+++ SakuraTranslate/Modules/Initialize.cs
@@ -1,0 +1,90 @@
+﻿using SimpleJSON;
+using System.Collections.Generic;
+using System.Net;
+using System;
+using XUnity.AutoTranslator.Plugin.Core.Endpoints;
+using XUnity.AutoTranslator.Plugin.Core.Utilities;
+using SakuraTranslate.Helpers;
+
+namespace SakuraTranslate
+{
+    public partial class SakuraTranslateEndpoint : ITranslateEndpoint
+    {
+        public void Initialize(IInitializationContext context)
+        {
+            _endpoint = context.GetOrCreateSetting<string>("Sakura", "Endpoint", "http://127.0.0.1:8080/v1/chat/completions");
+            _modelName = context.GetOrCreateSetting<string>("Sakura", "ModelName", "Sakura");
+            _modelVersion = context.GetOrCreateSetting<string>("Sakura", "ModelVersion", "1.0");
+            _modelType = TranslateModelHelper.GetTranslationModel(_modelName, _modelVersion);
+            if (!int.TryParse(context.GetOrCreateSetting<string>("Sakura", "MaxConcurrency", "1"), out _maxConcurrency))
+            {
+                _maxConcurrency = 1;
+            }
+            if (_maxConcurrency > ServicePointManager.DefaultConnectionLimit)
+            {
+                ServicePointManager.DefaultConnectionLimit = _maxConcurrency;
+            }
+            if (!bool.TryParse(context.GetOrCreateSetting<string>("Sakura", "UseDict", "False"), out _useDict))
+            {
+                _useDict = false;
+            }
+            _dictMode = context.GetOrCreateSetting<string>("Sakura", "DictMode", "Partial");
+            var dictStr = context.GetOrCreateSetting<string>("Sakura", "Dict", string.Empty);
+            if (string.IsNullOrEmpty(dictStr))
+            {
+                _useDict = false;
+                _fullDictStr = string.Empty;
+            }
+            else
+            {
+                try
+                {
+                    _dict = new Dictionary<string, List<string>>();
+                    var dictJObj = JSON.Parse(dictStr);
+                    foreach (var item in dictJObj)
+                    {
+                        try
+                        {
+                            var vArr = JSON.Parse(item.Value.ToString()).AsArray;
+                            List<string> vList;
+                            if (vArr.Count <= 0)
+                            {
+                                throw new Exception();
+                            }
+                            else if (vArr.Count == 1)
+                            {
+                                vList = new List<string> { JsonHelper.Unescape(vArr[0].ToString().Trim('\"')), string.Empty };
+                            }
+                            else
+                            {
+                                vList = new List<string> { JsonHelper.Unescape(vArr[0].ToString().Trim('\"')),
+                                    JsonHelper.Unescape(vArr[1].ToString().Trim('\"')) };
+                            }
+                            _dict.Add(JsonHelper.Unescape(item.Key.Trim('\"')), vList);
+                        }
+                        catch
+                        {
+                            _dict.Add(JsonHelper.Unescape(item.Key.Trim('\"')),
+                                new List<string> { JsonHelper.Unescape(item.Value.ToString().Trim('\"')), string.Empty });
+                        }
+                    }
+                    if (_dict.Count == 0)
+                    {
+                        _useDict = false;
+                        _fullDictStr = string.Empty;
+                    }
+                    else
+                    {
+                        var dictStrings = DictHelper.GetDictStringList(_dict);
+                        _fullDictStr = string.Join("\n", dictStrings.ToArray());
+                    }
+                }
+                catch
+                {
+                    _useDict = false;
+                    _fullDictStr = string.Empty;
+                }
+            }
+        }
+    }
+}
